right editor refactor to properly encapsulate editors and handlers
